+++ R/LinguGeo-Package.R
@@ -13,15 +13,3 @@ NULL
 #' @description A csv with long, lat and the spelling of "hunde" as well as classified spelling by phenomen.
 #' @format \code{"utils::read.csv"}
 NULL
-#' @docType data
-#' @name hunde_utm.shp - data
-#' @title Example shp Dataset
-#' @description A shp projected in UTM32N with long, lat and the spelling of "hunde" as well as classified spelling by phenomen.
-#' @format \code{"rgdal::readOGR"}
-NULL
-#' @docType data
-#' @name hunde_wgs.shp - data
-#' @title Example shp Dataset
-#' @description A csv projected in WGS84 with long, lat and the spelling of "hunde" as well as classified spelling by phenomen.
-#' @format \code{"rgdal::readOGR"}
-NULL